requirements.txt
numpy
torch
torchmetrics
pyyaml
rich
pandas